Thu, Jun 23, 2022, 5:55 AM +00:00

+++ filter-less/helpers.c
@@ -78,7 +78,7 @@ void blur(int height, int width, RGBTRIPLE image[height][width])
     int blurRed;
     int blurGreen;
     int blurBlue;
-    int counter;
+    float counter;
     for (int i = 0; i < height; i++)
     {
         for (int j = 0; j < width; j++)
